+++ core/index.js
@@ -1,20 +1,10 @@
 import Symbiote from './Symbiote.js';
 
-// Core modules:
 export { Symbiote };
 export { html } from './html.js';
 export { css } from './css.js';
 export { PubSub } from './PubSub.js';
 export { DICT } from './dictionary.js';
 export { animateOut } from './animateOut.js';
-
-// Utils:
-export { UID } from '../utils/UID.js';
-export { setNestedProp } from '../utils/setNestedProp.js';
-export { applyStyles } from '../utils/dom-helpers.js';
-export { applyAttributes } from '../utils/dom-helpers.js';
-export { create } from '../utils/dom-helpers.js';
-export { kebabToCamel } from '../utils/kebabToCamel.js';
-export { reassignDictionary } from '../utils/reassignDictionary.js';
 
 export default Symbiote;