+++ utils/index.js
@@ -1,0 +1,5 @@
+export { UID } from './UID.js';
+export { setNestedProp } from './setNestedProp.js';
+export { applyStyles, applyAttributes, create } from './dom-helpers.js';
+export { kebabToCamel } from './kebabToCamel.js';
+export { reassignDictionary } from './reassignDictionary.js';